Componentise "param" display.
this gives us a nice standalone place to implement param-related logic.
